Vegetation-plots from Kenya

A subset of http://www.givd.info/ID/AF-00-006SWEA-Dataveg including five references providing plots collected in Kenya.

Usage

Kenya_veg

Format

An object of class vegtable.

Source

http://www.givd.info/ID/AF-00-006

References

Bronner G (1990). Vegetation and land use in the Mathews Range area, Samburu-District, Kenya. J. Cramer, Berlin.

Bussmann RW (1994). The forests of Mount Kenya – vegetation, ecology, destruction and management of a tropical mountain forest ecosystem. Universität Bayreuth.

Bussmann RW (2002). Islands in the desert – forest vegetation of Kenya's smaller mountains and highland areas. Journal of East African Natural History 91: 27–79.

Fujiwara K, Furukawa T, Kiboi SK, Mathenge S, Mutiso P, Hayashi H, Meguro S (2014). Forest types and biodiversity around the Great Rift Valley in Kenya. Contributii Botanice 49: 143–178.

Schmitt K (1991). The vegetation of the Aberdare National Park Kenya. Universitätsverlag Wagner, Innsbruck.

Author

Miguel Alvarez kamapu78@gmail.com and Michael Curran currmi01@gmail.com

Examples

summary(Kenya_veg)
#> ## Metadata 
#>    db_name: Sweadataveg
#>    sp_list: Easplist
#>    dictionary: Swea
#>    object size: 9501 Kb 
#>    validity: TRUE 
#> 
#> ## Content 
#>    number of plots: 1946 
#>    plots with records: 1946 
#>    variables in header: 34 
#>    number of relations: 3 
#> 
#> ## Taxonomic List 
#>    taxon names: 3164 
#>    taxon concepts: 2392 
#>    validity: TRUE 
#>