Clean orphaned records in vegtable object

Delete entries in slots header and species orphaned by manipulation of slots.

Orphaned records generated by modifications in some slots may cause a loss on the validity of vegtable objects. This function should be applied to optimise the allocated size of a vegtable object, as well. Since running cleaning only once does not assure the deletion of all orphaned entries, it is recommended to run it at least twice. This repetition of cleaning is controlled by the argument times.

Usage

clean_once(object)

# S4 method for class 'vegtable'
clean(object, times = 2, ...)

Arguments

object

A vegtable object.

times

Numeric value indicating how many times should be the cleaning be repeated.

...

Further arguments passed from or to other methods.

Value

A clean vegtable object.

Author

Miguel Alvarez kamapu78@gmail.com

Examples

## Create an invalid object
veg <- Kenya_veg
veg@header <- veg@header[1:10, ]

## Resolve invalidity
veg <- clean(veg)