Merge concepts

Merge taxon concepts form into single ones or insert accepted names to slot samples.

This method is applied to a function defined in the package taxlist-package and only modify the slot species in the input object.

The use of taxa2samples() with merge_to argument will produce a similar result as using merge_taxa with level argument, but taxa2samples() will replace the records in slot samples by the respective accepted names without any modification in slot species. Additionally taxon concept IDs will be addes as columns in samples and taxon traits if indicated in argument add_traits.

Usage

# S4 method for vegtable,numeric,missing
merge_taxa(object, concepts, level, ...)

# S4 method for vegtable,missing,character
merge_taxa(object, concepts, level, ...)

taxa2samples(object, ...)

# S4 method for vegtable
taxa2samples(object, merge_to, include_levels, na.rm = FALSE, add_traits, ...)

Arguments

object

Object of class vegtable.

concepts

Numeric (integer) vector including taxon concepts to be merged.

level, merge_to

Character value indicating the level to which the taxa have to be merged.

...

Further arguments passed to taxlist::merge_taxa().

include_levels

Character vector indicating the levels to be considered in the output object. It can be used to exclude some taxonomic ranks.

na.rm

Logical value. Apply to records with missing information on taxonomic rank (i.e. for undetermined specimens).

add_traits

A character vector indicating variables in the slot taxonTraits to be added in slot samples.

Value

An object of class vegtable.

Author

Miguel Alvarez kamapu78@gmail.com

Examples

## Merge Olea capensis into one
summary(subset(Kenya_veg@species, grepl("Olea capensis", TaxonName),
  slot = "names"
), "all")
#> ------------------------------ 
#> concept ID: 50235 
#> view ID: 1 
#> level: subspecies 
#> parent: 52041 Olea capensis L. 
#> 
#> # accepted name: 
#> 50235 Olea capensis ssp. hochstetteri (Baker) Friis & P. S. Green 
#> 
#> # synonyms (1): 
#> 51708 Olea hochstetteri Baker 
#> ------------------------------ 
#> concept ID: 50432 
#> view ID: 1 
#> level: subspecies 
#> parent: 52041 Olea capensis L. 
#> 
#> # accepted name: 
#> 50432 Olea capensis ssp. welwitschii (Knobl. ) Friis & P. S. Green 
#> ------------------------------ 
#> concept ID: 52041 
#> view ID: 1 
#> level: species 
#> parent: none 
#> 
#> # accepted name: 
#> 52041 Olea capensis L. 
#> ------------------------------ 
Kenya_veg <- merge_taxa(Kenya_veg, c(52041, 50432, 50235))

## Check Olea capensis again
summary(subset(Kenya_veg@species, grepl("Olea capensis", TaxonName),
  slot = "names"
), "all")
#> ------------------------------ 
#> concept ID: 52041 
#> view ID: 1 
#> level: species 
#> parent: none 
#> 
#> # accepted name: 
#> 52041 Olea capensis L. 
#> 
#> # synonyms (3): 
#> 50235 Olea capensis ssp. hochstetteri (Baker) Friis & P. S. Green 
#> 51708 Olea hochstetteri Baker 
#> 50432 Olea capensis ssp. welwitschii (Knobl. ) Friis & P. S. Green 
#> ------------------------------ 

## Effect of taxa2samples by counting taxa
count_taxa(Kenya_veg, level = "genus")
#> [1] 90

Kenya_veg <- taxa2samples(Kenya_veg, merge_to = "genus")
count_taxa(Kenya_veg, level = "genus")
#> [1] 699