Retrieve synonyms or taxon concepts used in a data set

Plots records are rather linked to plant names than plant taxon concepts and used_synonyms() lists all synonyms linked to records in a vegtable object, including their respective accepted names.

On the other side, the function used_concepts() produces a subset of the taxonomic list embeded in the slot species including only taxonomic concepts linked to records in the slot samples.

Usage

used_synonyms(x, ...)

# S3 method for vegtable
used_synonyms(x, ...)

used_concepts(x, ...)

# S3 method for vegtable
used_concepts(x, keep_children = FALSE, keep_parents = FALSE, ...)

Arguments

x

A vegtable object.

...

Further arguments to be passed from or to another methods.

keep_children

A logical argument indicating whether children of selected taxa should be included in the output or not. This argument passed to get_children().

keep_parents

A logical argument indicating whether parents of selected taxa should be included in the output or not. This argument passed to get_parents().

Value

The function used_synonyms() returns a data frame including following variables:

SynonymID

ID of the taxon usage name applied as synonym.

Synonym

The synonym itself.

SynonymAuthor

Author of synonym.

TaxonConceptID

ID of the respective taxon concept.

AcceptedNameID

ID of the taxon usage name set as accepted name of the taxon concept.

AcceptedName

The respective accepted name.

AcceptedNameAuthor

The author of the accepted name.

The function used_concepts() returns a taxlist object including only taxa occurring in the plot observations of the input vegtable

object.

See also

Author

Miguel Alvarez kamapu78@gmail.com

Examples

## Synonyms used in the Kenya_veg
Synonyms <- used_synonyms(Kenya_veg)
head(Synonyms)
#>   SynonymID                                  Synonym   SynonymAuthor
#> 1     50168                  Clerodendron johnstonii              NA
#> 2     50222 Cyperus sesquiflorus ssp. appendiculatus (K. Schum.) Lye
#> 3     50277                           Teclea nobilis          Delile
#> 4     50297               Polystichum fuscopaleaceum          Alston
#> 5     50304                        Dombeya goetzenii       K. Schum.
#> 6     50225                 Solanum sessilistellatum          Bitter
#>   TaxonConceptID AcceptedNameID            AcceptedName AcceptedNameAuthor
#> 1          50167          50167 Clerodendrum johnstonii              Oliv.
#> 2          50221          50221  Cyperus afrosylvestris                Lye
#> 3          50276          50276          Vepris nobilis    (Delile) Mziray
#> 4          50296          50296   Polystichum discretum    (D. Don) J. Sm.
#> 5          50303          50303         Dombeya torrida (J.F. Gmel.) Bamps
#> 6          50224          50224  Solanum nigriviolaceum             Bitter

## Subset species list to used concepts
species <- used_concepts(Kenya_veg)
Kenya_veg@species
#> object size: 554 Kb 
#> validation of 'taxlist' object: TRUE 
#> 
#> number of taxon usage names: 3164 
#> number of taxon concepts: 2392 
#> trait entries: 102 
#> number of trait variables: 1 
#> taxon views: 3 
#> 
#> concepts with parents: 2237 
#> concepts with children: 957 
#> 
#> hierarchical levels: form < variety < subspecies < species < complex < genus < family 
#> number of concepts in level form: 0
#> number of concepts in level variety: 65
#> number of concepts in level subspecies: 52
#> number of concepts in level species: 1422
#> number of concepts in level complex: 0
#> number of concepts in level genus: 699
#> number of concepts in level family: 154
#> 
species
#> object size: 395.8 Kb 
#> validation of 'taxlist' object: TRUE 
#> 
#> number of taxon usage names: 2339 
#> number of taxon concepts: 1573 
#> trait entries: 97 
#> number of trait variables: 1 
#> taxon views: 3 
#> 
#> concepts with parents: 514 
#> concepts with children: 138 
#> 
#> hierarchical levels: form < variety < subspecies < species < complex < genus < family 
#> number of concepts in level form: 0
#> number of concepts in level variety: 65
#> number of concepts in level subspecies: 52
#> number of concepts in level species: 1366
#> number of concepts in level complex: 0
#> number of concepts in level genus: 90
#> number of concepts in level family: 0
#>