Retrieve synonyms or taxon concepts used in a data set

Plots records are rather linked to plant names than plant taxon concepts. The function used_synonyms() provides a quick report about synonyms used in a data set (a vegtable object) and their respective accepted names.

Additionally, not all taxon concepts included in the taxonomic list (slot species) may be recorded in the plot observations. In that case the function used_concepts() will optimize the size of the taxonomic list by discarding taxa that are not "in use". Alternatively parents or children of these taxa may be included in the output data set.

Usage

used_synonyms(x, ...)

# S4 method for vegtable
used_synonyms(x, ...)

used_concepts(x, ...)

# S4 method for vegtable
used_concepts(x, keep_children = FALSE, keep_parents = FALSE, ...)

Arguments

x

A vegtable object.

...

Further arguments to be passed from or to another methods.

keep_children

Argument passed to taxlist::get_children().

keep_parents

Argument passed to taxlist::get_parents().

Value

The function used_synonyms() returns a data frame including following variables:

SynonymID

ID of the taxon usage name applied as synonym.

Synonym

The synonym itself.

SynonymAuthor

Author of synonym.

TaxonConceptID

ID of the respective taxon concept.

AcceptedNameID

ID of the taxon usage name set as accepted name of the taxon concept.

AcceptedName

The respective accepted name.

AcceptedNameAuthor

The author of the accepted name.

The function used_concepts() returns a taxlist object including only taxa occurring in the plot observations of the input vegtable

object.

See also

Author

Miguel Alvarez kamapu78@gmail.com

Examples

## Synonyms used in the Kenya_veg
Synonyms <- used_synonyms(Kenya_veg)
head(Synonyms)
#>   SynonymID                                  Synonym   SynonymAuthor
#> 1     50168                  Clerodendron johnstonii              NA
#> 2     50222 Cyperus sesquiflorus ssp. appendiculatus (K. Schum.) Lye
#> 3     50277                           Teclea nobilis          Delile
#> 4     50297               Polystichum fuscopaleaceum          Alston
#> 5     50304                        Dombeya goetzenii       K. Schum.
#> 6     50225                 Solanum sessilistellatum          Bitter
#>   TaxonConceptID AcceptedNameID            AcceptedName AcceptedNameAuthor
#> 1          50167          50167 Clerodendrum johnstonii              Oliv.
#> 2          50221          50221  Cyperus afrosylvestris                Lye
#> 3          50276          50276          Vepris nobilis    (Delile) Mziray
#> 4          50296          50296   Polystichum discretum    (D. Don) J. Sm.
#> 5          50303          50303         Dombeya torrida (J.F. Gmel.) Bamps
#> 6          50224          50224  Solanum nigriviolaceum             Bitter